backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dateutil==2.8.2
google-cloud-firestore==2.16.0
google-cloud-storage==2.14.0
python-dotenv==1.0.1
